# Reuse memory for ELBO calculation