@sethaxen 2022-06-09
Thanks @cscherrer for the analysis and PR! I had considered reusing `u` before but ultimately decided not to in the interest of allowing the user to inspect all intermediates (they can currently access these draws if they desire). I'm not certain how useful that is in the end though.

The main reason I considered reusing `u` was due to memory. For very large models, say those with 10^6+ parameters, if Pathfinder runs 10^3 iterations, then we've already stored O(10^10) parameters. Drawing 5 draws from each distribution gives us 10^10 new parameters. So we're looking at worst at a doubling of the memory requirements by not reusing `u`. And it's not clear to me how big of a problem that is.