@sethaxen 2022-06-10
Keeping in mind #15, the goal is ultimately to allow the user to provide an object that configures an optimization procedure over the trace of distributions, with the default being `ELBOMaximization` or something like that. Such an object could have a type parameter that configures whether the samples drawn are kept or overwritten. So I'll take a stab at designing that interface. Then it should be pretty easy to finish up this PR based on that.